refactor packages for room db
