+++ app/src/main/java/com/mobicob/mobile/app/db/entity/Role.java
@@ -1,15 +1,19 @@
 package com.mobicob.mobile.app.db.entity;
 
+import android.arch.persistence.room.Entity;
+import android.arch.persistence.room.PrimaryKey;
+import android.support.annotation.NonNull;
+
+@Entity
 public class Role {
+
+    @PrimaryKey
+    @NonNull
     private int id;
     private String code;
     private String name;
 
     public Role(){}
-
-    public Role(int id){
-        this.id = id;
-    }
 
     public int getId() {
         return id;
